Implement stop subprocess execution from server
Made it so that now, when the server script receives a message from websockets asking for a subprocess termination, it will now connect as a client this time, with sockets, to the target subprocess server to send a "stop" message to it.

This allows for a StreamDeck-(websocket request)->Server-(socket message)->Subprocess communication chain. The server really serves as a launcher to start subprocesses, which would be impossible directly from the StreamDeck, which only has implementations to send WS request. (Unless I code a plugin for it, and GL with that)

+++ shop_watcher.py
@@ -1,5 +1,7 @@
 import asyncio
+import atexit
 import logging
+import time
 
 import cv2 as cv
 import mss
@@ -7,6 +9,12 @@ import numpy as np
 import websockets
 from skimage.metrics import structural_similarity as ssim
 from websockets import WebSocketException
+
+import denied_slots_db_handler as denied_sdh
+import single_instance
+import slots_db_handler as sdh
+import terminal_window_manager_v3 as twm_v3
+import constants
 
 # Configuration
 logging.basicConfig(level=logging.DEBUG,
@@ -18,10 +26,19 @@ SCREEN_CAPTURE_AREA = {"left": 1883, "top": 50, "width": 37, "height": 35}
 TEMPLATE_IMAGE_PATH = 'opencv/dota_shop_top_right_icon.jpg'
 WEBSOCKET_URL = "ws://127.0.0.1:8080/"
 SECONDARY_WINDOW_NAMES = ['opencv_shop_watcher']
+SCRIPT_NAME = "dota2_shop_watcher"
 
-secondary_window_spawned = asyncio.Event()
-mute_print_feedback = asyncio.Event()
+secondary_windows_have_spawned = asyncio.Event()
+mute_main_loop_print_feedback = asyncio.Event()
 stop_loop = asyncio.Event()
+
+
+def exit_countdown():
+    """Give a bit of time to read terminal exit statements"""
+    for seconds in reversed(range(1, 6)):
+        print("\r" + f'cmd will close in {seconds} seconds...', end="\r")
+        time.sleep(1)
+    exit()
 
 
 async def capture_window(area):
@@ -54,15 +71,15 @@ async def establish_ws_connection():
 
 
 async def handle_socket_client(reader, writer):
-    addr = writer.get_extra_info('peername')
-    print(f"Socket server connected by {addr}")
     while True:
         data = await reader.read(1024)
         if not data:
             print("Socket client disconnected")
             break
         message = data.decode()
-        print(f"Received from subprocess: {message}")
+        if message == constants.STOP_SUBPROCESS_MESSAGE:
+            stop_loop.set()
+        print(f"Received: {message}")
         writer.write(b"ACK from WebSocket server")
         await writer.drain()
     writer.close()
@@ -74,8 +91,14 @@ async def run_socket_server():
     addr = server.sockets[0].getsockname()
     print(f"Serving on {addr}")
 
-    async with server:
+    try:
         await server.serve_forever()
+    except asyncio.CancelledError:
+        print("Socket server task was cancelled. Stopping server")
+    finally:
+        server.close()
+        await server.wait_closed()
+        print("Server closed")
 
 
 async def scan_for_shop_and_notify(ws):
@@ -87,12 +110,12 @@ async def scan_for_shop_and_notify(ws):
         gray_frame = cv.cvtColor(frame, cv.COLOR_BGR2GRAY)
         match_value = await compare_images(gray_frame, template)
         cv.imshow(SECONDARY_WINDOW_NAMES[0], gray_frame)
-        secondary_window_spawned.set()
+        secondary_windows_have_spawned.set()
 
         if cv.waitKey(1) == ord("q"):
             break
-        if not mute_print_feedback.is_set():
-            print(f"SSIM: {match_value}", end="\r")
+        if not mute_main_loop_print_feedback.is_set():
+            print(f"SSIM: {match_value:.6f}", end="\r")
 
         if match_value >= 0.8 and not shop_is_currently_open:
             shop_is_currently_open = True
@@ -110,22 +133,50 @@ async def scan_for_shop_and_notify(ws):
     if ws:
         await ws.close()
         cv.destroyAllWindows()
+    print('loop terminated')
 
 
 async def main():
-    socket_server = run_socket_server()
-    ws = None
-    try:
-        await asyncio.gather(socket_server)
-        ws = await establish_ws_connection()
-        await scan_for_shop_and_notify(ws)
-    except KeyboardInterrupt:
-        print("KeyboardInterrupt")
-    finally:
-        if ws:
-            await ws.close()
-        cv.destroyAllWindows()
+    """If there are no single instance lock file, start the Dota2 shop_watcher
+     module. At launch, reposition immediately the terminal providing feedback
+    regarding its execution. Shortly after, reposition the secondary window
+    which the module spawns. This is all done in an asynchronous way thanks
+    to a database providing information used for the window positions"""
+    if single_instance.lock_exists():
+        slot = twm_v3.handle_window(twm_v3.WindowType.DENIED_SCRIPT,
+                                    SCRIPT_NAME)
+        atexit.register(denied_sdh.free_slot, slot)
+        print("\n>>> Lock file is present: exiting... <<<")
+
+    else:
+        slot = twm_v3.handle_window(twm_v3.WindowType.ACCEPTED_SCRIPT,
+                                    SCRIPT_NAME, SECONDARY_WINDOW_NAMES)
+
+        single_instance.create_lock_file()
+        atexit.register(single_instance.remove_lock)
+        atexit.register(sdh.free_slot_named, SCRIPT_NAME)
+        socket_server_task = asyncio.create_task(run_socket_server())
+        mute_main_loop_print_feedback.set()
+
+        ws = None
+        try:
+            ws = await establish_ws_connection()
+            main_task = asyncio.create_task(scan_for_shop_and_notify(ws))
+            await secondary_windows_have_spawned.wait()
+            twm_v3.adjust_secondary_windows(slot, SECONDARY_WINDOW_NAMES)
+            mute_main_loop_print_feedback.clear()
+            await main_task
+        except KeyboardInterrupt:
+            print("KeyboardInterrupt")
+        finally:
+            socket_server_task.cancel()
+            await socket_server_task
+            cv.destroyAllWindows()
+            if ws:
+                await ws.close()
 
 
 if __name__ == "__main__":
+    logging.info("START")
     asyncio.run(main())
+    input('enter to q')
